+++ project-editor/src/renderer/components/Modals/SettingsTabs/SettingsTabMain.vue
@@ -1,0 +1,6 @@
+<template>
+    <div>
+        <h1>Main</h1>
+        <h2>Стас Арбуз</h2>
+    </div>
+</template>