+++ project-editor/src/renderer/components/Modals/SettingsTabs/SettingsTabOther.vue
@@ -1,0 +1,6 @@
+<template>
+    <div>
+        <h1>Other</h1>
+        <p>Стас, да ты заебал со своим арбузом!</p>
+    </div>
+</template>